Moved AgencyType enum into `data/constants` module
This is to break the dependency some of the test data
modules and `server/data/prisonApi` which depends on
`bunyan`.

This is to prevent bunyan being bundled in the
integration tests/browser JS where Node' `util` (used
by `bunyan`) is no longer available by default as of
Cypress 15.x.
